+++ src/routes/graphql/interfaces/post.ts
@@ -1,0 +1,6 @@
+export interface IPost {
+  id: string;
+  title: string;
+  content: string;
+  authorId: string;
+}
