+++ src/routes/graphql/interfaces/profile.ts
@@ -1,0 +1,7 @@
+export interface IProfile {
+  id: string;
+  isMale: boolean;
+  yearOfBirth: number;
+  userId: string;
+  memberTypeId: string;
+}
