+++ src/routes/graphql/types/memberTypeId.ts
@@ -1,0 +1,13 @@
+import { GraphQLEnumType } from 'graphql';
+
+export const MemberTypeId = new GraphQLEnumType({
+  name: "MemberTypeId",
+  values: {
+    basic: { 
+      value: "basic"
+    },
+    business: {
+      value: "business"
+    },
+  }
+});
